requirements.txt
h5py==2.7.0
Keras==2.0.5
python-midi>=0.2.3
numpy==1.13.0
tensorflow==1.2.0
